requirements.txt
streamlit
finance-datareader
tensorflow
scikit-learn
pandas
numpy
matplotlib
seaborn
langchain
langchain-openai
python-dotenv
scipy
openai
